xenserver-mcp/pyproject.toml
[project]
name = "xenserver-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for XenServer devices"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp",
    "XenAPI",
    "python-dotenv",
    "starlette",
    "uvicorn",
]

[dependency-groups]
dev = []

[project.urls]
Homepage = "https://github.com/xenserver-rahm/xenserver-mcp"
Repository = "https://github.com/xenserver-rahm/xenserver-mcp.git"